shell done

+++ scripts/run.sh
@@ -1,15 +1,22 @@
 #!/bin/bash
 cur_path=$(pwd)
-
-BINPATH=/home/kouushou/Github/spmv/bin
-cd ${BINPATH};
-rm -rf run_sh_temp
-./spmv $1 8 8 > run_sh_temp
-
-if [[ $? != 0 ]]
+last=$(cat lastMtx.info)
+echo $1 > lastMtx.info
+if [[ "$1" > "$last" ]];
 then
-  echo $1 >> ${cur_path}/err_mtx.txt
+  echo $1 "in running"
+  BINPATH=/home/kouushou/Github/spmv/bin
+  cd ${BINPATH};
+  rm -rf run_sh_temp
+  ./spmv $1 8 8 > run_sh_temp
+
+  if [[ $? != 0 ]]
+  then
+    echo $1 >> ${cur_path}/err_mtx.txt
+  else
+    cat run_sh_temp >> ${cur_path}/$2
+  fi
+  rm -rf run_sh_temp
 else
-  cat run_sh_temp >> ${cur_path}/$2
+  echo $1 "has runned"
 fi
-rm -rf run_sh_temp
